refactor: 재사용 위해 모달 컴포넌트 리팩토링
